Set up Authentication
When user successfully logs in, we return a token
that gets added to local storage in the client's browser.

When a user tries to perform an action restricted to admin's
only (edit, delete, add), we pass in that token
in the headers of our fetch calls.

If it's invalid then nothing can happen. Else we perform that
action.